Add the dom manipulation operations to a separatefile
