+++ src/domManip.js
@@ -1,0 +1,31 @@
+  let cityName = document.getElementById('cityName');
+  let search = document.getElementById('citySearch');
+  let checkBox = document.getElementById('toggle-checkbox');
+
+  let city_name = document.getElementById('city-name');
+  let city_initial = document.getElementById('city-code');
+  let weather_description = document.getElementById('description');
+  let weather_ico = document.getElementById('weather-ico');
+  let temperature = document.getElementById('temperature');
+  let current_day = document.getElementById('day');
+
+  let feel_like = document.getElementById('feels_like');
+  let min_temp = document.getElementById('min_temp');
+  let max_temp = document.getElementById('max_temp');
+  let humidity = document.getElementById('humidity');
+
+  export {
+      cityName,
+      search,
+      checkBox,
+      city_name,
+      city_initial,
+      weather_description,
+      weather_ico,
+      temperature,
+      current_day,
+      feel_like,
+      min_temp,
+      max_temp,
+      humidity
+  }
